requirements.txt
Flask==3.0.0
pinecone-client==3.0.0
pinecone-datasets==0.7.0